refactor: additional refactoring
